+++ client/src/presentation/components/layout/contact-form.js
@@ -1,7 +1,7 @@
-const contactForm = (title, id) => {
+const contactForm = (title, formId, sendId, handler) => {
   const form = document.createElement("form");
   form.className = "modal-form";
-  form.id = id;
+  form.id = formId;
   form.action = "#";
   form.innerHTML = `
   <h1>${title}</h1>
@@ -12,7 +12,7 @@ const contactForm = (title, id) => {
   </div>
   <div class="form-control">
     <label for="phone-number">Phone number</label><br />
-    <input type="number" name="phone-number" /><br />
+    <input type="number" name="phone" /><br />
     <small> Error message </small>
   </div>
   <div class="form-control">
@@ -21,12 +21,30 @@ const contactForm = (title, id) => {
     <small> Error message </small>
   </div>
   <div class="form-control">
+  <label for="subject">Subject</label><br />
+  <input type="text" name="subject" /><br />
+  <small> Error message </small>
+</div>
+  <div class="form-control">
     <label for="Message">Message</label><br />
     <textarea rows="14" cols="60" required> </textarea>
     <small> Error message </small>
   </div>
-  <button id = "send-${id}-form" class="button form-button">Send</button>
    `;
+  const divEl = document.createElement("div");
+  const button = document.createElement("button");
+  button.id = sendId;
+  button.classList.add("button", "form-button");
+  button.type = "submit";
+  button.innerHTML = "Send";
+  button.addEventListener("click", handler);
+  const br = document.createElement("br");
+  const span = document.createElement("span");
+  span.id = "status";
+  divEl.appendChild(button);
+  divEl.appendChild(br);
+  divEl.appendChild(span);
+  form.appendChild(divEl);
   return form;
 };
